Fix for unw_word_t on 32-bit
git-svn-id: http://parabix.costar.sfu.ca/svn/icGREP/icgrep-devel@5502 a82c6ad6-7e1c-4be0-8e68-26c050b43eaf

+++ icgrep/IR_Gen/CBuilder.cpp
@@ -850,7 +850,7 @@ void CBuilder::__CreateAssert(Value * const assertion, StringRef failureMessage)
             if (pc == 0) {
                 break;
             }
-            stack.push_back(static_cast<uint64_t>(pc));
+            stack.push_back(pc);
         }
         #elif defined(HAS_EXECINFO)
         for (;;) {
